+++ selfdrive/ui/ui_state.py
@@ -2,7 +2,7 @@ import pyray as rl
 from enum import Enum
 from cereal import messaging, log
 from openpilot.common.params import Params, UnknownKeyName
-from openpilot.selfdrive.ui.prime_state import PrimeState
+from openpilot.selfdrive.ui.lib.prime_state import PrimeState
 
 UI_BORDER_SIZE = 30
